refactor: extract behavioral CAPTCHA logic into dedicated middleware #544
